# openapi.json
{
  "components": {
    "schemas": {
...
      "Body_upload_questionnaire_api_questionnaires_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_questionnaire_api_questionnaires_post",
        "type": "object"
      },
...
      "ErrorOut": {
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorOut",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ItemOut": {
        "properties": {
          "code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Code"
          },
          "csf_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Csf Id"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "position": {
            "title": "Position",
            "type": "integer"
          },
          "question": {
            "title": "Question",
            "type": "string"
          },
          "row_ref": {
            "title": "Row Ref",
            "type": "string"
          },
          "topic": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic"
          }
        },
        "required": [
          "id",
          "position",
          "row_ref",
          "code",
          "topic",
          "question",
          "csf_id"
        ],
        "title": "ItemOut",
        "type": "object"
      },
...
      "Mapping": {
        "additionalProperties": false,
        "description": "Where the questionnaire lives in its file. Columns are letters for xlsx and csv alike (A is the first\ncolumn); header_row is 1-based. topic_col is optional: without it, the last section row is the topic.",
        "properties": {
          "answer_col": {
            "pattern": "^[A-Z]{1,3}$",
            "title": "Answer Col",
            "type": "string"
          },
          "comments_col": {
            "anyOf": [
              {
                "pattern": "^[A-Z]{1,3}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comments Col"
          },
          "header_row": {
            "maximum": 1000.0,
            "minimum": 1.0,
            "title": "Header Row",
            "type": "integer"
          },
          "id_col": {
            "anyOf": [
              {
                "pattern": "^[A-Z]{1,3}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id Col"
          },
          "question_col": {
            "pattern": "^[A-Z]{1,3}$",
            "title": "Question Col",
            "type": "string"
          },
          "scope": {
            "anyOf": [
              {
                "maxLength": 32,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scope"
          },
          "sheet": {
            "anyOf": [
              {
                "maxLength": 31,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sheet"
          },
          "topic_col": {
            "anyOf": [
              {
                "pattern": "^[A-Z]{1,3}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic Col"
          }
        },
        "required": [
          "sheet",
          "header_row",
          "id_col",
          "question_col",
          "answer_col",
          "comments_col"
        ],
        "title": "Mapping",
        "type": "object"
      },
...
      "PreviewRow": {
        "properties": {
          "answer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Answer"
          },
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "question": {
            "title": "Question",
            "type": "string"
          },
          "row": {
            "title": "Row",
            "type": "integer"
          },
          "topic": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic"
          }
        },
        "required": [
          "row",
          "id",
          "question",
          "topic",
          "answer"
        ],
        "title": "PreviewRow",
        "type": "object"
      },
...
      "QuestionnaireDetail": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "detected": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Mapping"
              },
              {
                "type": "null"
              }
            ]
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "format": {
            "enum": [
              "xlsx",
              "csv",
              "builtin"
            ],
            "title": "Format",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "item_count": {
            "title": "Item Count",
            "type": "integer"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ItemOut"
            },
            "title": "Items",
            "type": "array"
          },
          "latest_run_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Run Id"
          },
          "mapping": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Mapping"
              },
              {
                "type": "null"
              }
            ]
          },
          "preview": {
            "items": {
              "$ref": "#/components/schemas/PreviewRow"
            },
            "title": "Preview",
            "type": "array"
          },
          "sheets": {
            "items": {
              "type": "string"
            },
            "title": "Sheets",
            "type": "array"
          },
          "source": {
            "enum": [
              "sample",
              "upload",
              "drive",
              "csf"
            ],
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "id",
          "filename",
          "source",
          "format",
          "sheets",
          "detected",
          "mapping",
          "preview",
          "item_count",
          "latest_run_id",
          "created_at",
          "items"
        ],
        "title": "QuestionnaireDetail",
        "type": "object"
      },
      "QuestionnaireOut": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "detected": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Mapping"
              },
              {
                "type": "null"
              }
            ]
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "format": {
            "enum": [
              "xlsx",
              "csv",
              "builtin"
            ],
            "title": "Format",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "item_count": {
            "title": "Item Count",
            "type": "integer"
          },
          "latest_run_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Run Id"
          },
          "mapping": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Mapping"
              },
              {
                "type": "null"
              }
            ]
          },
          "preview": {
            "items": {
              "$ref": "#/components/schemas/PreviewRow"
            },
            "title": "Preview",
            "type": "array"
          },
          "sheets": {
            "items": {
              "type": "string"
            },
            "title": "Sheets",
            "type": "array"
          },
          "source": {
            "enum": [
              "sample",
              "upload",
              "drive",
              "csf"
            ],
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "id",
          "filename",
          "source",
          "format",
          "sheets",
          "detected",
          "mapping",
          "preview",
          "item_count",
          "latest_run_id",
          "created_at"
        ],
        "title": "QuestionnaireOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "VART",
    "version": "2.0.0.dev0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/questionnaires": {
      "get": {
        "description": "Every questionnaire in the workspace, newest first (a built-in `csf` one included, Plan 6B).",
        "operationId": "list_questionnaires_api_questionnaires_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/QuestionnaireOut"
                  },
                  "title": "Response List Questionnaires Api Questionnaires Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "A state-changing request from another site (POST, PUT, PATCH, DELETE)"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unknown id, another workspace's id, or the workspace is gone"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "The action conflicts with the item's state"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "A per-network limit or the model budget; see Retry-After"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "The demo is full, or model calls are off"
          }
        },
        "summary": "List Questionnaires",
        "tags": [
          "questionnaires"
        ]
      },
      "post": {
        "description": "Multipart xlsx or csv. Answers the detected mapping and a preview; no items exist until the visitor\nconfirms with PUT .../mapping. 422 for a refused file, a file over 1 MB (`MAX_QUESTIONNAIRE_BYTES`), or a\nworkspace that already holds 5 questionnaires (`MAX_QUESTIONNAIRES`); 429 per network (`upload`); 503 when\nthe demo is full. The sheet names are stored at upload, so listing never re-parses the file.",
        "operationId": "upload_questionnaire_api_questionnaires_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_questionnaire_api_questionnaires_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionnaireOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "A state-changing request from another site (POST, PUT, PATCH, DELETE)"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unknown id, another workspace's id, or the workspace is gone"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "The action conflicts with the item's state"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ErrorOut"
                    },
                    {
                      "$ref": "#/components/schemas/HTTPValidationError"
                    }
                  ]
                }
              }
            },
            "description": "Refused input: a sentence, or FastAPI's validation list"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "A per-network limit or the model budget; see Retry-After"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "The demo is full, or model calls are off"
          }
        },
        "summary": "Upload Questionnaire",
        "tags": [
          "questionnaires"
        ]
      }
    },
    "/api/questionnaires/sample/{name}": {
      "post": {
        "description": "`vsq-a` (xlsx) or `mvsp-b` (csv), mapped and itemised at once. 404 for another name. Idempotent:\nwhen the workspace already has that sample, it is answered again and nothing is stored. A new one counts\nunder the per-network `upload` limit (429), the storage breaker (503) and `MAX_QUESTIONNAIRES` (422).",
        "operationId": "load_sample_questionnaire_api_questionnaires_sample__name__post",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionnaireDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "A state-changing request from another site (POST, PUT, PATCH, DELETE)"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unknown id, another workspace's id, or the workspace is gone"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "The action conflicts with the item's state"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ErrorOut"
                    },
                    {
                      "$ref": "#/components/schemas/HTTPValidationError"
                    }
                  ]
                }
              }
            },
            "description": "Refused input: a sentence, or FastAPI's validation list"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "A per-network limit or the model budget; see Retry-After"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "The demo is full, or model calls are off"
          }
        },
        "summary": "Load Sample Questionnaire",
        "tags": [
          "questionnaires"
        ]
      }
    },
    "/api/questionnaires/{questionnaire_id}": {
      "delete": {
        "description": "Remove a questionnaire and its items, so the 5-per-workspace cap does not force a reset. 409 when a run\nused it (reset the workspace to start over).",
        "operationId": "delete_questionnaire_api_questionnaires__questionnaire_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "questionnaire_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Questionnaire Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "A state-changing request from another site (POST, PUT, PATCH, DELETE)"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unknown id, another workspace's id, or the workspace is gone"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "The action conflicts with the item's state"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "A per-network limit or the model budget; see Retry-After"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "The demo is full, or model calls are off"
          }
        },
        "summary": "Delete Questionnaire",
        "tags": [
          "questionnaires"
        ]
      },
      "get": {
        "operationId": "get_questionnaire_api_questionnaires__questionnaire_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "questionnaire_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Questionnaire Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionnaireDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "A state-changing request from another site (POST, PUT, PATCH, DELETE)"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unknown id, another workspace's id, or the workspace is gone"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "The action conflicts with the item's state"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "A per-network limit or the model budget; see Retry-After"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "The demo is full, or model calls are off"
          }
        },
        "summary": "Get Questionnaire",
        "tags": [
          "questionnaires"
        ]
      }
    },
...
  }
}